+++ ntex/CHANGES.md
@@ -1,5 +1,14 @@
 # Changes
 
+## [0.4.9] - 2021-11-20
+
+* Update rustls to 0.20
+* Update webpki to 0.22
+* Update webpki-roots to 0.22
+* Update tokio-rustls to 0.23
+* Update tokio-ssl to 0.6.3
+* Adapt code for rustls breaking changes
+
 ## [0.4.8] - 2021-11-08
 
 * Add Clone impl for connect::ConnectError